Modified JSONUtils.getExpandedJSON() to be JSONObject.expand()
