+++ src/test/resources/JSONObjWithArraysTestIn2_2.json
