+++ src/test/resources/JSONObjWithArraysTestIn2_3.json
