+++ src/test/resources/JSONObjWithArraysTestOut1.json
