+++ src/test/resources/JSONObjWithArraysTestOut2.json
