+++ src/test/resources/JSONObjWithArraysTestOut2_1_1.json
