+++ src/test/resources/JSONObjWithArraysTestOut2_1_2.json
