+++ src/test/resources/JSONObjWithArraysTestOut2_1_3.json
