+++ src/test/resources/JSONObjWithArraysTestOut2_2_2.json
