+++ src/test/resources/JSONObjWithArraysTestOut2_3_1.json
