+++ src/test/resources/JSONObjWithArraysTestOut2_3_2.json
