+++ src/test/resources/JSONObjWithArraysTestOut3.json
